Change preference key for medical history pull timestamp
When moving from api v1 to api v2, the structure of the pull token was changed from a timestamp to a base64 encoded string. When an app update happens, the older timestamp that was saved on the device cannot be sent to the v2 api. This commit changes the preference key for the timestamp and ignores the previous one, so that sync happens without issues.

(cherry picked from commit 0c83ee02f758bb6373a3d74b81a6ce63c976c7f5)