Add file StateSizeLogger to the flow package
This will be responsible for logging the sizes of all the different
states when the activity is destroyed. The reason this is in the `flow`
package is because it needs access to `flow.KeyManager` which is a
package scoped class.

(cherry picked from commit 5636d599047c5f3c5dd93a7f7f05a4950f1d1e87)