Remove loading of overdue item count on the home screen (#1716)
This is being done because the overdue query is quite complex and we
load both the count of items and a paginated list of the items when
we open the open screen which causes a slowdown in future, immediately
required database queries.

This commit disables showing the overdue count temporarily until we
figure out how to improve the app performance.

(cherry picked from commit 68c6babaa28cf446ed3484586887a4ef944255fd)